apps/backend/pyproject.toml
[project]
name = "querymind-engine"
version = "1.0.0"

dependencies = [
    "pandas==2.3.3",
    "pyarrow==23.0.1",
    "pandas-gbq==0.33.0",
    "langgraph==1.0.8",
    "langchain-core==1.2.13",
    "langchain-google-genai==4.2.0",
    "google-cloud-bigquery==3.40.1",
    "google-cloud-bigquery-storage==2.36.1",
    "python-dotenv==1.2.1",
]

[tool.setuptools]
packages = ["core", "engines"]

[build-system]
requires = ["setuptools>=82.0.0"]
build-backend = "setuptools.build_meta"
